Added search facets support and added minimal terms facet objects

+++ jres/src/main/java/com/blacklocus/jres/model/search/Hits.java
@@ -24,14 +24,14 @@ import java.util.List;
 */
 public class Hits {
 
-    private Integer total;
+    private Long total;
 
     @JsonProperty("max_score")
     private Double maxScore;
 
     private List<Hit> hits;
 
-    public Integer getTotal() {
+    public Long getTotal() {
         return total;
     }
